setting up workspace v1
