+++ controllers/contacts/helpers.js
@@ -1,0 +1,9 @@
+import { Contact } from "#models/Contact.js";
+
+export const listContacts = async () => {
+  try {
+    return await Contact.find();
+  } catch (error) {
+    return error;
+  }
+};
