+++ models/Contact.js
@@ -1,0 +1,4 @@
+import mongoose from "mongoose";
+import { contactSchema } from "#validators/contactSchema.js";
+
+export const Contact = mongoose.model("contact", contactSchema, "contacts");
