--- models/contacts.json
@@ -1,1 +1,0 @@
-[{"id":"AeHIrLTr6JkxGE6SN-0Rw","name":"Bartek","email":"bartek@test.pl","phone":432456231},{"id":"qdggE76Jtbfd9eWJHrssH","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"drsAJ4SHPYqZeG-83QTVW","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"vza2RIzNGIwutCVCs4mCL","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"05olLMgyVQdWRwgKfg5J6","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"1DEXoP8AuCGYc1YgoQ6hw","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"Z5sbDlS7pCzNsnAHLtDJd","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"C9sjBfCo4UJCWjzBnOtxl","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"e6ywwRe4jcqxXfCZOj_1e","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"jg9VhvwM03KAMFCBO5RkZ","name":"Bartek","email":"bartek@test.pl","phone":423432423},{"id":"s9pcyc__PNjfPMjwrriI4","name":"Bartek","email":"bartek@test.pl","phone":423432423},{"id":"iVi4D6zI32_jLisToWNDr","name":"Bartek","email":"bartek@test.pl","phone":"423432423"},{"id":"wBDhpqKKZY3OH-HesvFcY","name":"Bartek","email":"bartek@test.pl","phone":423432423},{"id":"WsV4YSVrGmVo3zE-oUKIz","name":"Bartek","email":"bartek@test.pl","phone":423432423},{"id":"hsnmJdC0l25jzRMlkj4lD","name":"Bartek","email":"bartek@test.pl","phone":432456231}]
